+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- **[Feature]** Add `ErasedScheduler` trait, a dyn-compatible version of `Scheduler`.
+
 # 0.1.0 (2025-06-04)
 
 - **[Feature]** First release.